Move stream utils to their own module
