+++ baremaps-cli/src/main/java/com/baremaps/cli/pipeline/Execute.java
@@ -19,8 +19,8 @@ import static com.baremaps.core.tile.VariableUtils.interpolate;
 import com.baremaps.blob.BlobStore;
 import com.baremaps.cli.Options;
 import com.baremaps.core.postgres.PostgresUtils;
-import com.baremaps.osm.stream.StreamException;
-import com.baremaps.osm.stream.StreamUtils;
+import com.baremaps.stream.StreamException;
+import com.baremaps.stream.StreamUtils;
 import com.google.common.base.Splitter;
 import java.net.URI;
 import java.nio.charset.StandardCharsets;
