+++ baremaps-core/src/main/java/com/baremaps/core/database/ImportService.java
@@ -14,8 +14,8 @@
 
 package com.baremaps.core.database;
 
-import static com.baremaps.osm.stream.ConsumerUtils.consumeThenReturn;
-import static com.baremaps.osm.stream.StreamUtils.batch;
+import static com.baremaps.stream.ConsumerUtils.consumeThenReturn;
+import static com.baremaps.stream.StreamUtils.batch;
 
 import com.baremaps.blob.Blob;
 import com.baremaps.blob.BlobStore;
