+++ baremaps-osm/pom.xml
@@ -18,12 +18,22 @@
     </dependency>
     <dependency>
       <groupId>com.baremaps</groupId>
+      <artifactId>baremaps-stream</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>com.baremaps</groupId>
       <artifactId>baremaps-testing</artifactId>
       <scope>test</scope>
     </dependency>
     <dependency>
       <groupId>org.apache.logging.log4j</groupId>
       <artifactId>log4j-slf4j-impl</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>com.baremaps</groupId>
+      <artifactId>baremaps-stream</artifactId>
+      <version>0.6.6-SNAPSHOT</version>
+      <scope>compile</scope>
     </dependency>
   </dependencies>
