+++ baremaps-osm/src/main/java/com/baremaps/osm/change/OsmChangeSpliterator.java
@@ -32,7 +32,7 @@ import com.baremaps.osm.domain.Member.MemberType;
 import com.baremaps.osm.domain.Node;
 import com.baremaps.osm.domain.Relation;
 import com.baremaps.osm.domain.Way;
-import com.baremaps.osm.stream.StreamException;
+import com.baremaps.stream.StreamException;
 import java.io.InputStream;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
