+++ baremaps-osm/src/main/java/com/baremaps/osm/function/BlockFunction.java
@@ -17,7 +17,7 @@ package com.baremaps.osm.function;
 import com.baremaps.osm.domain.Block;
 import com.baremaps.osm.domain.DataBlock;
 import com.baremaps.osm.domain.HeaderBlock;
-import com.baremaps.osm.stream.StreamException;
+import com.baremaps.stream.StreamException;
 import java.util.function.Function;
 
 /**
