+++ baremaps-osm/src/main/java/com/baremaps/osm/pbf/HeaderBlockReader.java
@@ -20,7 +20,7 @@ import com.baremaps.osm.domain.Blob;
 import com.baremaps.osm.domain.Bound;
 import com.baremaps.osm.domain.Header;
 import com.baremaps.osm.domain.HeaderBlock;
-import com.baremaps.osm.stream.StreamException;
+import com.baremaps.stream.StreamException;
 import com.google.protobuf.InvalidProtocolBufferException;
 import java.time.LocalDateTime;
 import java.time.ZoneOffset;
