+++ baremaps-osm/src/main/java/com/baremaps/osm/xml/OsmXmlSpliterator.java
@@ -31,7 +31,7 @@ import com.baremaps.osm.domain.Member.MemberType;
 import com.baremaps.osm.domain.Node;
 import com.baremaps.osm.domain.Relation;
 import com.baremaps.osm.domain.Way;
-import com.baremaps.osm.stream.StreamException;
+import com.baremaps.stream.StreamException;
 import com.google.common.primitives.Ints;
 import com.google.common.primitives.Longs;
 import java.io.InputStream;
