+++ baremaps-osm/src/test/java/com/baremaps/osm/xml/XmlChangeSpliteratorTest.java
@@ -21,8 +21,8 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 import com.baremaps.osm.change.OsmChangeSpliterator;
 import com.baremaps.osm.domain.Change;
-import com.baremaps.osm.stream.AccumulatingConsumer;
-import com.baremaps.osm.stream.HoldingConsumer;
+import com.baremaps.stream.AccumulatingConsumer;
+import com.baremaps.stream.HoldingConsumer;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Spliterator;
