+++ baremaps-stream/pom.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+
+    <modelVersion>4.0.0</modelVersion>
+    <artifactId>baremaps-stream</artifactId>
+    <name>baremaps-stream</name>
+
+    <parent>
+        <artifactId>baremaps</artifactId>
+        <groupId>com.baremaps</groupId>
+        <version>0.6.6-SNAPSHOT</version>
+    </parent>
+
+</project>
