+++ baremaps-stream/src/main/java/com/baremaps/stream/AccumulatingConsumer.java
@@ -12,31 +12,33 @@
  * the License.
  */
 
-package com.baremaps.osm.stream;
+package com.baremaps.stream;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.function.Consumer;
 
 /**
- * A {@code Consumer} that holds the latest value it accepted.
+ * A {@code Consumer} that accumulates the values it accepts.
  *
  * @param <T>
  */
-public class HoldingConsumer<T> implements Consumer<T> {
+public class AccumulatingConsumer<T> implements Consumer<T> {
 
-  private T value;
+  private final List<T> values = new ArrayList<>();
 
-  /** {@inheritDoc} */
+  /** Stores the accepted value. */
   @Override
   public void accept(T value) {
-    this.value = value;
+    values.add(value);
   }
 
   /**
-   * Returns the holded value.
+   * Returns the accumulated values.
    *
-   * @return the holded value.
+   * @return the accumulated values.
    */
-  public T value() {
-    return value;
+  public List<T> values() {
+    return values;
   }
 }
