+++ baremaps-stream/src/main/java/com/baremaps/stream/ConsumerUtils.java
@@ -12,33 +12,27 @@
  * the License.
  */
 
-package com.baremaps.osm.stream;
+package com.baremaps.stream;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.function.Consumer;
+import java.util.function.Function;
 
-/**
- * A {@code Consumer} that accumulates the values it accepts.
- *
- * @param <T>
- */
-public class AccumulatingConsumer<T> implements Consumer<T> {
+/** Utility methods for dealing with consumers. */
+public class ConsumerUtils {
 
-  private final List<T> values = new ArrayList<>();
-
-  /** Stores the accepted value. */
-  @Override
-  public void accept(T value) {
-    values.add(value);
-  }
+  private ConsumerUtils() {}
 
   /**
-   * Returns the accumulated values.
+   * Transforms a consumer into a function.
    *
-   * @return the accumulated values.
+   * @param consumer the consumer
+   * @param <T> the type
+   * @return the function
    */
-  public List<T> values() {
-    return values;
+  public static <T> Function<T, T> consumeThenReturn(Consumer<T> consumer) {
+    return t -> {
+      consumer.accept(t);
+      return t;
+    };
   }
 }
