+++ baremaps-stream/src/main/java/com/baremaps/stream/StreamException.java
@@ -12,7 +12,7 @@
  * the License.
  */
 
-package com.baremaps.osm.stream;
+package com.baremaps.stream;
 
 /**
  * When a checked exception occurs in a stream, it is a good practice to wrap that exception in an
